lsm works with runfile
