+++ lsm/fortran/funcphys.F
