+++ lsm/physcons.py
