+++ lsm/python/physcons.py
