+++ lsm/python/stencils_dimK.py
