+++ lsm/python/stencils_dimK2.py
