+++ lsm/update.sh
